fix: Respect highlight's `linenums` config for `pycon` examples in docstrings
The reasoning is that someone enabling line numbers globally supposedly also wants them for code blocks in docstrings. There's in my opinion no difference between a code block in a Markdown document and a code block in a docstring, they both serve documentation purposes and should be subject to the same rendering rules.

If that makes someone unhappy, we can always consider adding an option to hide line numbers for code blocks in docstrings.

Related-to-#192: https://github.com/mkdocstrings/python/issues/192